backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
requests==2.32.4